<execute> Changed filePrefix to a UUID
This is because we're starting to use the `parallel` command.
When using `parallel`, the execute command will run into each other's
scope because the a timestamp is taken as a fileprefix. Using a UUID is more
secure because, it's theoritically impossible to generate the same UUID twice,
espacially on such a small scope